requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
accelerate=1.3.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.35.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.4.0=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.12.14=hbcca054_0
cbor=1.0.0=pypi_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.0=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.12.1=pypi_0
datasets=3.2.0=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.13.2=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
docker-pycreds=0.4.0=pypi_0
docstring-parser=0.16=pypi_0
einops=0.8.0=pypi_0
email-validator=2.1.2=pypi_0
eval-type-backport=0.2.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
filelock=3.15.1=pyhd8ed1ab_0
flagembedding=1.3.4=pypi_0
flash-attn=2.3.4=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.4.1=pypi_0
fschat=0.2.35=pypi_0
fsspec=2023.10.0=pypi_0
fuzzywuzzy=0.18.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.3.0=h59595ed_1
gmpy2=2.1.5=py39h048c657_1
gpustat=1.1.1=pypi_0
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.28.1=pypi_0
idna=3.7=pypi_0
ijson=3.3.0=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.5=pypi_0
inscriptis=2.5.3=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
ipykernel=6.29.4=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.5=pypi_0
ir-datasets=0.5.9=pypi_0
isal=1.6.1=pypi_0
jedi=0.19.1=pypi_0
jieba=0.42.1=pypi_0
jinja2=3.1.4=pyhd8ed1ab_0
jiter=0.6.0=pypi_0
joblib=1.4.2=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
ld_impl_linux-64=2.40=hf3520f5_6
levenshtein=0.26.0=pypi_0
libabseil=20240116.2=cxx17_he02047a_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.0.4=0
libcurand=10.3.6.39=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h77fa898_10
libgomp=13.2.0=h77fa898_10
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libprotobuf=4.25.3=h08a7969_0
libsentencepiece=0.2.0=hc656cff_3
libsqlite=3.46.0=hde9e2c9_0
libstdcxx-ng=13.2.0=hc0a3c3a_10
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=h4ab18f5_1
llvm-openmp=15.0.7=h0cdce71_0
lxml=5.3.1=pypi_0
lz4=4.4.3=pypi_0
markdown-it-py=3.0.0=pypi_0
markdown2=2.4.13=pypi_0
markupsafe=2.1.5=py39hd1e30aa_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mkl=2022.1.0=hc2b9512_224
mpc=1.3.1=hfe3b2da_0
mpfr=4.2.1=h9458935_1
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.5=h59595ed_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pyhd8ed1ab_0
nh3=0.2.17=pypi_0
ninja=1.11.1.1=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-ml-py=12.555.43=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.62.0=pypi_0
openssl=3.3.1=h4ab18f5_0
orjson=3.10.5=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.4=pypi_0
peft=0.11.1=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.2.0=pypi_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=5.27.1=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=16.1.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11=2.12.0=pypi_0
pydantic=1.10.16=pypi_0
pydantic-core=2.18.4=pypi_0
pygments=2.18.0=pypi_0
pynvml=11.5.0=pypi_0
python=3.9.19=h0755675_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-levenshtein=0.26.0=pypi_0
python-multipart=0.0.9=pypi_0
python_abi=3.9=4_cp39
pytorch=2.1.1=py3.9_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py39hd1e30aa_1
pyzmq=26.0.3=pypi_0
rapidfuzz=3.9.7=pypi_0
readline=8.2=h8228510_1
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rouge=1.0.1=pypi_0
rouge-score=0.1.2=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
sentence-transformers=3.4.1=pypi_0
sentencepiece=0.2.0=hf3d152e_3
sentencepiece-python=0.2.0=py39h9839300_3
sentencepiece-spm=0.2.0=hc656cff_3
sentry-sdk=2.5.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.0=pypi_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
shtab=1.7.1=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.37.2=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.12=pypyh9d50eac_103
tensor-parallel=2.0.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.19.0=pypi_0
torchaudio=2.1.1+cu121=pypi_0
torchtriton=2.1.0=py39
torchvision=0.16.1+cu121=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.44.2=pypi_0
trec-car-tools=2.6=pypi_0
trl=0.9.4=pypi_0
typer=0.12.3=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tyro=0.8.4=pypi_0
tzdata=2024.1=pypi_0
ujson=5.10.0=pypi_0
unlzw3=0.2.3=pypi_0
unzip=6.0=h611a1e1_0
urllib3=2.2.1=pypi_0
uvicorn=0.30.1=pypi_0
uvloop=0.19.0=pypi_0
wandb=0.17.1=pypi_0
warc3-wet=0.2.5=pypi_0
warc3-wet-clueweb09=0.2.5=pypi_0
watchfiles=0.22.0=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.13=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
widgetsnbextension=4.0.13=pypi_0
xopen=2.0.2=pypi_0
xxhash=3.4.1=pypi_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
yarl=1.9.4=pypi_0
zipp=3.19.2=pypi_0
zlib-ng=0.4.3=pypi_0
zlib-state=0.1.9=pypi_0
